Add `UpdateConfig` method to multisig
A successful `UpdateConfig` proposal will modify the config for the
Multisig.

Adds additional verification to proposals which ensure that
`UpdateConfig` messages that are part of a proposal will not set a
threshold that is unreachable. This prevents a proposal from being
added which will not be possible to execute due to an invalid updated
config.